New Follower PM Tweak
Tweaked the new follower PM code to validate the user and display_name
of the new follower. This tweak is added because for reasons I've not
been able to work out every now and again the system finds the following
follower "_total":144,"_links":{"self":"
https://api.twitch.tv/kraken/channels/n_skid11/follows?direction=DESC&limit=2&offset=0".
This tweak should prevent this by making sure the user and display name
only contain valid charactures for them.