Added Dice Roller
Removed debug messages for IRC send command, because it would print the
oauth to the console.
Disabled TwitchAPI intergreation, they've changed how it's handle and
now it's broken, need recoding.
Added the ability for the bot to make dice rolls for the users.
Added the ability to assign a GM so that you can PM a dice to the GM.