MySQLHandler Update
Updated the MySQLHandler code to the updated version I use at work,
should fix a bug that could cause the code to be unable to reconnect to
the MySQL server. Also increases the buffer size for MySQL statements,
and allows a request for the effected number of rows of a query.